+++ src/Stack/Exec.hs
@@ -26,6 +26,14 @@ defaultEnvSettings = EnvSettings
     , esStackExe = True
     }
 
+-- | Environment settings which do not embellish the environment
+plainEnvSettings :: EnvSettings
+plainEnvSettings = EnvSettings
+    { esIncludeLocals = False
+    , esIncludeGhcPackagePath = False
+    , esStackExe = False
+    }
+
 -- | Execute a process within the Stack configured environment.
 exec :: (HasConfig r, MonadReader r m, MonadIO m, MonadLogger m, MonadThrow m)
         => EnvSettings -> String -> [String] -> m b